Added a date picker and formatting error checks for the date inputs

+++ app/src/main/java/com/example/librarybookapp/view/BookInfoScreen.kt
@@ -13,7 +13,11 @@ import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
@@ -23,24 +27,27 @@ import kotlinx.coroutines.launch
 @Composable
 fun BookInfoScreen(bookListViewModel: BookListViewModel, onDelete: () -> Unit, onEdit: () -> Unit) {
     val currentBook = bookListViewModel.bookInfo.value
+    var showDialog by remember { mutableStateOf(false) }
     val coroutineScope = rememberCoroutineScope()
+    val pad = 16.dp
+
     Column(modifier = Modifier.fillMaxSize().padding(top = 30.dp)) {
-        Text(text = "Book Title: ")
+        Text(text = "Book Title: ", modifier = Modifier.padding(start = pad))
         BookInfoCard(currentBook?.bookTitle ?: "")
 
-        Text(text = "Book Author: ")
+        Text(text = "Book Author: ", modifier = Modifier.padding(start = pad))
         BookInfoCard(currentBook?.bookAuthor ?: "")
 
-        Text(text = "Book Genre: ")
+        Text(text = "Book Genre: ", modifier = Modifier.padding(start = pad))
         BookInfoCard(currentBook?.bookGenre ?: "")
 
-        Text(text = "Date Published: ")
-        BookInfoCard(currentBook?.datePublished ?: "")
+        Text(text = "Date Published: ", modifier = Modifier.padding(start = pad))
+        BookInfoCard((currentBook?.datePublished ?: "").toString())
 
-        Text(text = "Pages: ")
-        BookInfoCard(currentBook?.readingProgress ?: "")
+        Text(text = "Pages: ", modifier = Modifier.padding(start = pad))
+        BookInfoCard((currentBook?.pages ?: "").toString())
 
-        Text(text = "Date Added:")
+        Text(text = "Date Added:", modifier = Modifier.padding(start = pad))
         BookInfoCard(currentBook?.dateAdded ?: "")
 
         Row(horizontalArrangement = Arrangement.SpaceEvenly,
@@ -51,22 +58,33 @@ fun BookInfoScreen(bookListViewModel: BookListViewModel, onDelete: () -> Unit, o
                     start = 10.dp),
                 colors = ButtonDefaults.buttonColors(Color(0xFF6650a4))
             ) {
-                Text(text = "Edit"
+                Text(text = "Edit",
+                    color = Color.White
                 )
             }
             Button(modifier = Modifier.padding(top = 10.dp)
                 ,onClick = {
-                coroutineScope.launch {
-                    bookListViewModel.deleteBook(currentBook!!.id)
-                }
-                onDelete()
+                    showDialog = true
             },
-                colors = ButtonDefaults.buttonColors(Color.Red)) {
+                colors = ButtonDefaults.buttonColors(Color.Red))
+            {
                 Row{
-                    Icon(Icons.Default.Delete, contentDescription = "Delete")
-                    Text(text = "Delete")
+                    Icon(Icons.Default.Delete, contentDescription = "Delete", tint = Color.White)
+                    Text(text = "Delete",
+                        color = Color.White)
                 }
             }
         }
-    }
+        if (showDialog) {
+            ConfirmationDialog(
+                onDismiss = { showDialog = false },
+                onConfirm = {
+                    showDialog = false
+                    coroutineScope.launch {
+                        bookListViewModel.deleteBook(currentBook!!.id)
+                    }
+                    onDelete()
+                })
+        }
+    } //Column End
 }
